+++ src/events/events.ts
@@ -1,1 +1,3 @@
 export type { MidSelectEvent } from './mid-select';
+export type { MidInvalidShowEvent } from './mid-invalid-show';
+export type { MidInvalidHideEvent } from './mid-invalid-hide';