+++ src/index.ts
@@ -23,5 +23,6 @@ export { MinidSpinner } from './components/spinner.component.ts';
 export { MinidStepIndicator } from './components/step-indicator.component.ts';
 export { MinidTextfield } from './components/textfield.component.ts';
 export { MinidTooltip } from './components/tooltip.component.ts';
+export { MinidValidationMessage } from './components/validation-message.component.ts';
 
 export * from './events/events.ts';