+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
-A changelog for logging changes
+A changelog for logging changes.
+
+## 0.0.106
+
+- New validation message component
+- Added invalid styling for phone input
+- Fixed bugs in phone input
 
 ## 0.0.105
 
@@ -13,7 +19,7 @@ A changelog for logging changes
   This command can be added to `package.json` for easy copying:
 
 ```bash
-    "copy-assets": "cp -aR node_modules/@felleslosninger/minid-elements/dist/icons/ ./public/icons && cp -aR node_modules/@felleslosninger/minid-elements/dist/flags/ ./public/flags"
+    cp -aR node_modules/@felleslosninger/minid-elements/dist/icons/ ./public/icons && cp -aR node_modules/@felleslosninger/minid-elements/dist/flags/ ./public/flags
 ```
 
 ## 0.0.103